Move getAssociatedMetricThreshold function in new file #2710
