+++ visualization/app/codeCharta/ui/artificialIntelligence/ArtificialIntelligenceHelper/util/getMetricThresholds.ts
@@ -1,0 +1,5 @@
+import { metricThresholdsByLanguage } from "../artificialIntelligence.metricThresholds"
+
+export function getAssociatedMetricThresholds(programmingLanguage: string) {
+	return programmingLanguage === "java" ? metricThresholdsByLanguage.java : metricThresholdsByLanguage.miscellaneous
+}
